# Conflicts:
#	src/main/java/Main.java

+++ src/main/java/EdgeDetection.java
@@ -2,18 +2,18 @@ import static java.lang.Math.*;
 
 public class EdgeDetection { // extends PixelMap {
     private static final int[][] laplacian_operator = {{-1,-1,-1},{-1,8,-1},{-1,-1,-1}};
-    public int[][] sharpnessValue;
-    public EdgeDetection(int[][][] input_image) {
+    public short[][] sharpnessValue;
+    public EdgeDetection(short[][][] input_image) {
         int height = input_image.length;
         int width = input_image[0].length;
-        sharpnessValue = new int[height][width];
-        int red;
-        int green;
-        int blue;
+        sharpnessValue = new short[height][width];
+        short red;
+        short green;
+        short blue;
         for(int i=0; i<height; i++) {
             for(int j=0; j<width; j++){
                 if(i>0 && j>0 && i<height-1 && j<width-1){
-                    int[][][] laplacian_matrix = {{input_image[i-1][j-1],input_image[i-1][j],input_image[i-1][j+1]},
+                    short[][][] laplacian_matrix = {{input_image[i-1][j-1],input_image[i-1][j],input_image[i-1][j+1]},
                             {input_image[i][j-1],input_image[i][j],input_image[i][j+1]},
                             {input_image[i+1][j-1],input_image[i+1][j],input_image[i+1][j+1]}};
 
@@ -25,19 +25,19 @@ public class EdgeDetection { // extends PixelMap {
                             sharpnessValue[i][j] += (red + green + blue) * laplacian_operator[k][l];
                         }
                     }
-                    sharpnessValue[i][j] = abs(sharpnessValue[i][j]);
+                    sharpnessValue[i][j] = (short) abs(sharpnessValue[i][j]);
                 }
                 else{
                     red = input_image[i][j][1];
                     green = input_image[i][j][2];
                     blue = input_image[i][j][3];
-                    sharpnessValue[i][j] = red + green + blue; //to trzeba będzie udoskonalić
+                    sharpnessValue[i][j] = (short) (red + green + blue); //to trzeba będzie udoskonalić
                 }
             }
         }
     }
 
-    public int[][] getFocusMap(){
+    public short[][] getFocusMap(){
         return this.sharpnessValue;
     }
 }
