+++ src/main/java/GaussianBlur.java
@@ -1,18 +1,18 @@
 class GaussianBlur {
-    private int[][][] image;
+    private short[][][] image;
     private final static int[][] weight_matrix = {{1,2,1},{2,4,2},{1,2,1}};
-    public GaussianBlur(int[][][] input_image) {
+    public GaussianBlur(short[][][] input_image) {
         int height = input_image.length;
         int width = input_image[0].length;
-        this.image = new int[height][width][4];
+        this.image = new short[height][width][4];
         for(int i=0; i<height; i++) {
             for(int j=0; j<width; j++){
                 if(i>0 && j>0 && i<height-1 && j<width-1){
-                    int[][][] blur_matrix = {{input_image[i-1][j-1],input_image[i-1][j],input_image[i-1][j+1]},
+                    short[][][] blur_matrix = {{input_image[i-1][j-1],input_image[i-1][j],input_image[i-1][j+1]},
                                             {input_image[i][j-1],input_image[i][j],input_image[i][j+1]},
                                             {input_image[i+1][j-1],input_image[i+1][j],input_image[i+1][j+1]}};
-                    int blurred_pixel=0;
-                    int green;
+                    short blurred_pixel=0;
+                    short green;
                     for(int k=0; k<3; k++){
                         for(int l=0; l<3; l++){
                             //oblicz średnią ważoną pixel.green
@@ -32,10 +32,10 @@ class GaussianBlur {
             }
         }
     }
-    public int[] getPixel(int x, int y){
+    public short[] getPixel(int x, int y){
         return this.image[x][y];
     }
-    public int[][][] getImage(){
+    public short[][][] getImage(){
         return this.image;
     }
 }
